Use an enum to represent suggested byte lengths
65c816 assembly does have immediate instructions which take different amount
of bytes depending on exact mode. It's fairly well recognized to use two or
four digit hexadecimal literals to force a particular mode for immediate
instruction, and it's not really worth enforcing size annotations for most
common case.

Of course, this is error prone, so there are plans to have a type checker
whose purpose is to check if correct immediate instruction for a given mode
is used. I think the language would be better of without this feature, but
the train already left with Asar, so a warning for wrong types is second
best thing. That said, no other patterns are going to be supported, unless
they turn out to be surprisingly popular (which I'm fairly sure will happen,
unfortunately).

Old version of this code stored number of hexadecimal characters, but that
model is error-prone, and not really working for binary literals which could
as well use this too.