Implement opcode parser
This implements absolute and indirect addressing mode. The reason why those
two particular modes were picked is that implementing indirect addressing
mode is fairly tricky, as you need to check if entire expression is in parens.

The solution I decided to go with is lookahead to check if there is token
ahead that would change the expression into a standard expression when testing
indirect addressing. If there is, this particular addressing mode is avoided.
This is somewhat of a hack, but hey.